pyproject.toml
[project]
name = "grok-jr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.1.2",
    "langchain-xai>=1.1.0",
    "langgraph>=1.0.4",
    "pyserial>=3.5",
    "python-dotenv>=1.2.1",
    "xai-sdk>=1.5.0",
]
